refactored friends functionality: split add/search into component and cleaned up manage friends page
